refactor(core): common:integration 모듈 제거 및 관련 코드 통합
- `common:integration` 모듈 삭제
- WebClient 설정 및 관련 코드를 `common:web-reactive`로 통합
- 불필요한 의존성 제거 및 패키지 경로 수정
- 서비스 빌드 파일 및 설정에서 `common:integration` 참조 제거
