--- common/integration/build.gradle.kts
@@ -1,34 +1,0 @@
-import org.springframework.boot.gradle.tasks.bundling.BootJar
-
-plugins {
-    alias(libs.plugins.kotlin.jvm)
-    alias(libs.plugins.kotlin.spring)
-    alias(libs.plugins.spring.boot)
-    alias(libs.plugins.spring.dependency.management)
-}
-
-java {
-    withSourcesJar()
-    withJavadocJar()
-}
-
-tasks.named<BootJar>("bootJar") {
-    enabled = false
-}
-
-tasks.named<Jar>("jar") {
-    enabled = true
-}
-
-dependencies {
-    api(libs.spring.boot.starter.webclient)
-    api(libs.jackson.module.kotlin)
-
-    implementation(project(":common:exception"))
-    implementation(project(":common:util"))
-    implementation(libs.kotlin.reflect)
-
-    testImplementation(libs.kotest.runner.junit5)
-    testImplementation(libs.kotest.assertions.core)
-    testImplementation(libs.mockk)
-}
