--- common/integration/src/main/kotlin/com/nextmall/common/integration/config/JacksonConfig.kt
@@ -1,11 +1,0 @@
-package com.nextmall.common.integration.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import tools.jackson.module.kotlin.KotlinModule
-
-@Configuration
-class JacksonConfig {
-    @Bean
-    fun kotlinModule(): KotlinModule = KotlinModule.Builder().build()
-}
