--- common/integration/src/main/kotlin/com/nextmall/common/integration/exception/ConnectionFailedException.kt
@@ -1,11 +1,0 @@
-package com.nextmall.common.integration.exception
-
-class ConnectionFailedException(
-    context: IntegrationErrorContext,
-    cause: Throwable? = null,
-) : IntegrationException(
-        errorCode = IntegrationErrorCode.CONNECTION_FAILED,
-        expected = false,
-        cause = cause,
-        context = context,
-    )
