--- common/integration/src/main/kotlin/com/nextmall/common/integration/exception/IntegrationErrorContext.kt
@@ -1,11 +1,0 @@
-package com.nextmall.common.integration.exception
-
-import com.nextmall.common.exception.ErrorResponse
-
-data class IntegrationErrorContext(
-    val url: String?,
-    val statusCode: Int?,
-    val responseBody: String?,
-    val errorResponse: ErrorResponse? = null,
-    val serviceName: String? = null,
-)
