--- common/integration/src/main/kotlin/com/nextmall/common/integration/exception/ServerErrorException.kt
@@ -1,11 +1,0 @@
-package com.nextmall.common.integration.exception
-
-class ServerErrorException(
-    context: IntegrationErrorContext,
-    cause: Throwable? = null,
-) : IntegrationException(
-        errorCode = IntegrationErrorCode.SERVER_ERROR,
-        expected = false,
-        cause = cause,
-        context = context,
-    )
