--- common/integration/src/main/kotlin/com/nextmall/common/integration/exception/TimeoutException.kt
@@ -1,11 +1,0 @@
-package com.nextmall.common.integration.exception
-
-class TimeoutException(
-    context: IntegrationErrorContext,
-    cause: Throwable? = null,
-) : IntegrationException(
-        errorCode = IntegrationErrorCode.TIMEOUT,
-        expected = false,
-        cause = cause,
-        context = context,
-    )
