+++ common/web-mvc/src/main/kotlin/com/nextmall/common/web/mvc/exception/GlobalExceptionHandlerSupport.kt
@@ -3,7 +3,7 @@ package com.nextmall.common.web.mvc.exception
 import com.nextmall.common.exception.ErrorResponse
 import com.nextmall.common.exception.base.BaseException
 import com.nextmall.common.exception.code.CommonErrorCode
-import com.nextmall.common.integration.exception.IntegrationException
+import com.nextmall.common.web.core.exception.IntegrationException
 import com.nextmall.common.web.core.exception.HttpStatusMapper
 import jakarta.validation.ConstraintViolationException
 import org.slf4j.MDC
