+++ services/bff-service/src/main/kotlin/com/nextmall/bff/client/user/WebClientUserServiceClient.kt
@@ -4,7 +4,7 @@ import com.nextmall.bff.client.user.request.CreateUserClientRequest
 import com.nextmall.bff.client.user.response.CreateUserClientResponse
 import com.nextmall.bff.client.user.response.UserViewClientResponse
 import com.nextmall.bff.security.PassportTokenPropagationFilter
-import com.nextmall.common.integration.support.WebClientFactory
+import com.nextmall.common.web.reactive.client.WebClientFactory
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.client.bodyToMono
 import reactor.core.publisher.Mono
